+++ app/src/main/java/io/itch/fr/quizgame/feature_start/presentation/StartPageViewModel.kt
@@ -1,28 +1,20 @@
 package io.itch.fr.quizgame.feature_start.presentation
 
-import dagger.hilt.android.lifecycle.HiltViewModel
-
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import io.itch.fr.quizgame.feature_start.domain.GetQuestionsUseCase
-import kotlinx.coroutines.launch
+import androidx.navigation.NavController
+import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
 class StartPageViewModel @Inject constructor(
-    private val getQuestionsUseCase: GetQuestionsUseCase
+    private val navController: NavController
 ) : ViewModel() {
+
     fun startQuiz() {
-        viewModelScope.launch {
-            // Fetch questions using the use case
-            val questions = getQuestionsUseCase()
-            // Start the quiz with the fetched questions
-            // (navigate to the Quiz Page with the questions)
-        }
+        navController.navigate("quiz")
     }
 
     fun viewHistory() {
-        // View history logic
-        // (navigate to the History Page)
+        navController.navigate("history")
     }
 }